Decompose widgets.ts into smaller modules for better ts-node performance
Split the large widgets.ts (~3400 lines) into smaller, focused modules:
- widgets/base.ts - Widget base class and shared interfaces
- widgets/containers.ts - Layout containers (VBox, HBox, Grid, etc.)
- widgets/inputs.ts - Input widgets (Button, Entry, Checkbox, etc.)
- widgets/display.ts - Display widgets (Label, Image, Table, etc.)
- widgets/canvas.ts - Canvas primitives (CanvasLine, CanvasCircle, etc.)
- widgets/index.ts - Re-exports all widgets

The original widgets.ts now re-exports from widgets/index.ts for
backwards compatibility with existing imports.

This decomposition improves ts-node compilation performance by allowing
incremental compilation of smaller files rather than one large monolith.
